post-install: improve kernel install procedure
previously, we left dracut to its own devices. this
would have resulted in a catastrophic failure to boot.
now, we feed dracut some options which ensure that it
generates a working, clean initramfs.

we also check for existing kernel/initramfs pairs
and move them to a new pair that GRUB will find and
can use for a fallback

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>