genstrap: Decompress kernel modules for livecd
kmod in the gentoo livecd does not support zstd compressed kernel
modules making it impossible to load modules.
This makes the generated livecd useable on non laptop devices. Most of
the problems could probably avoided by adding the appropiate modules to
the initramfs but I falied to do that.
This change is however useful to support devices connected after boot.

Signed-off-by: Janne Grunau <j@jannau.net>