genstrap: reduce the size of the root image
The LiveCD has grown slightly, and no longer fits in the ramdisk
we set up. Rather than increase the size of the ramdisk, we can
instead refrain from decompressing the kernel modules and take
more useless firmware out to save space. This is doubly useful
since we are space constrained on the ESP, and the reference
distro may at some point switch to an ESP-as-boot schema.

Fixes: #6

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>