Adding Asahi Fedora compatibility
I had to change the lines from "pacman -S ..." to "dnf install -y ..." to gain compatibility for the new Asahi Fedora Remix which is standard for the Asahi Installer now.

Also the grub folder is called grub2 in Fedora and for Dracut to call dmsquash-live the parameter "hostonly" needs to get set to "no".