update dracut configuration and update files to match new kernel config
In https://github.com/chadmed/asahi-overlay/pull/59
we propose a new kernel configuration. This has
implications for how dracut needs to be configured,
as we build more of the Asahi drivers directly into the
kernel than the previous configuration did.

While we are editing the dracut configuration, we also pull in a couple of updates from the upstream asahi-scripts that the files come from.